Moved RdaService, Fixed Javadoc warnings in Handwritten java code
