+++ apps/bfd-model/bfd-model-rda/src/main/java/gov/cms/bfd/model/rda/RdaApiProgress.java
@@ -26,8 +26,11 @@ import lombok.experimental.FieldNameConstants;
 @FieldNameConstants
 @Table(name = "rda_api_progress", schema = "rda")
 public class RdaApiProgress {
+  /** Represents the enum ClaimType. */
   public enum ClaimType {
+    /** FISS. */
     FISS,
+    /** MCS. */
     MCS
   }
