+++ apps/bfd-pipeline/bfd-pipeline-rda-grpc/src/main/java/gov/cms/bfd/pipeline/rda/grpc/server/RdaJsonMessageSourceFactory.java
@@ -1,6 +1,7 @@
 package gov.cms.bfd.pipeline.rda.grpc.server;
 
 import com.google.common.io.CharSource;
+import gov.cms.bfd.pipeline.rda.grpc.server.service.RdaService;
 import gov.cms.mpsm.rda.v1.FissClaimChange;
 import gov.cms.mpsm.rda.v1.McsClaimChange;
 import lombok.AllArgsConstructor;
