+++ apps/bfd-pipeline/bfd-pipeline-rda-grpc/src/main/java/gov/cms/bfd/pipeline/rda/grpc/server/RdaServer.java
@@ -3,6 +3,7 @@ package gov.cms.bfd.pipeline.rda.grpc.server;
 import com.google.common.base.Strings;
 import gov.cms.bfd.pipeline.rda.grpc.ThrowableAction;
 import gov.cms.bfd.pipeline.rda.grpc.ThrowableConsumer;
+import gov.cms.bfd.pipeline.rda.grpc.server.service.RdaService;
 import io.grpc.ManagedChannel;
 import io.grpc.Server;
 import io.grpc.ServerBuilder;
