+++ apps/bfd-pipeline/bfd-pipeline-rda-grpc/src/main/java/gov/cms/bfd/pipeline/rda/grpc/server/service/RdaService.java
@@ -1,7 +1,9 @@
-package gov.cms.bfd.pipeline.rda.grpc.server;
+package gov.cms.bfd.pipeline.rda.grpc.server.service;
 
 import com.google.common.annotations.VisibleForTesting;
 import com.google.protobuf.Empty;
+import gov.cms.bfd.pipeline.rda.grpc.server.MessageSource;
+import gov.cms.bfd.pipeline.rda.grpc.server.RdaMessageSourceFactory;
 import gov.cms.mpsm.rda.v1.ApiVersion;
 import gov.cms.mpsm.rda.v1.ClaimRequest;
 import gov.cms.mpsm.rda.v1.FissClaimChange;
