+++ apps/bfd-pipeline/bfd-pipeline-rda-grpc/src/test/java/gov/cms/bfd/pipeline/rda/grpc/server/service/RdaServiceTest.java
@@ -1,9 +1,11 @@
-package gov.cms.bfd.pipeline.rda.grpc.server;
+package gov.cms.bfd.pipeline.rda.grpc.server.service;
 
 import static org.junit.jupiter.api.Assertions.assertEquals;
 import static org.mockito.Mockito.*;
 
 import com.google.protobuf.Empty;
+import gov.cms.bfd.pipeline.rda.grpc.server.MessageSource;
+import gov.cms.bfd.pipeline.rda.grpc.server.RdaMessageSourceFactory;
 import gov.cms.mpsm.rda.v1.ApiVersion;
 import gov.cms.mpsm.rda.v1.ClaimRequest;
 import gov.cms.mpsm.rda.v1.FissClaimChange;
